Move timeline page so its own file and tidy up css imports in the timeline page plugin
